SM1 - Restructured the Project
* Updated the Folder Structure to Accomodate Multiple Layout Designs
