+++ lib/screens/Login/login.dart
@@ -1,39 +1,36 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_doctor_ui/main.dart';
 import 'package:http/http.dart' as http;
-import 'package:intl/intl.dart';
-import 'package:flutter_doctor_ui/screens/login.dart';
-import 'package:flutter_doctor_ui/widgets/CustomTextField.dart';
-import 'package:flutter_doctor_ui/widgets/DatePickerField.dart';
-import 'package:flutter_doctor_ui/widgets/PasswordField.dart';
-import 'package:flutter_doctor_ui/widgets/PrimaryButton.dart';
-import 'package:flutter_doctor_ui/widgets/SignInFacebookButton.dart';
-import 'package:flutter_doctor_ui/widgets/SignInGoogleButton.dart';
+import 'package:flutter_doctor_ui/screens/Dashboard/dashboard.dart';
+import 'package:flutter_doctor_ui/screens/Registration/registration.dart';
+import 'package:flutter_doctor_ui/widgets/General/CustomTextField.dart';
+import 'package:flutter_doctor_ui/widgets/General/PasswordField.dart';
+import 'package:flutter_doctor_ui/widgets/General/PrimaryButton.dart';
+import 'package:flutter_doctor_ui/widgets/General/SignInFacebookButton.dart';
+import 'package:flutter_doctor_ui/widgets/General/SignInGoogleButton.dart';
 
-class Registration extends StatefulWidget {
-  static String routeName = "/registration";
-  const Registration({super.key});
+class Login extends StatefulWidget {
+  static String routeName = "/login";
+  const Login({super.key});
 
   @override
-  State<Registration> createState() => _RegistrationState();
+  State<Login> createState() => _LoginState();
 }
 
-class _RegistrationState extends State<Registration> {
-  final TextEditingController emailController = TextEditingController();
-  final TextEditingController phoneController = TextEditingController();
+class _LoginState extends State<Login> {
+  final TextEditingController inputController = TextEditingController();
   final TextEditingController passwordController = TextEditingController();
-  final TextEditingController birthdateController = TextEditingController();
-
   bool obscurePassword = true;
 
-  _RegisterPatient() async {
+  int? parsedData;
+
+  Future<int> _LoginPatient() async {
     String url =
-        'http://10.0.2.2:8080/flutter-mobile-backend-ui/index.php/patient/registerPatient';
+        'http://10.0.2.2:8080/flutter-mobile-backend-ui/index.php/patient/loginPatient';
 
     Map<String, String> data = {
-      'email': emailController.text,
-      'phone': phoneController.text,
+      'input': inputController.text,
       'password': passwordController.text,
-      'birthdate': birthdateController.text,
     };
 
     try {
@@ -42,36 +39,26 @@ class _RegistrationState extends State<Registration> {
         body: data,
       );
 
-      // Handle the response as needed
+      // Handle the response
       print('Response Status Code: ${response.statusCode}');
       print('Response Body: ${response.body}');
+
+      // Check the status code and parse data accordingly
+      if (response.statusCode == 200) {
+        // Successful response, parse integer or other data
+        int parsedData = int.tryParse(response.body) ?? 0;
+        return parsedData;
+      } else {
+        // Handle other status codes or error responses
+        print('Error: ${response.statusCode}');
+        // Return a default value or throw an exception if needed
+        return 0;
+      }
     } catch (error) {
       // Handle errors
-      print(error);
-    }
-  }
-
-  Future<void> _selectDate(BuildContext context) async {
-    final DateTime? picked = await showDatePicker(
-      context: context,
-      initialDate: DateTime.now(),
-      firstDate: DateTime(1900),
-      lastDate: DateTime.now(),
-      builder: (BuildContext context, Widget? child) {
-        return Theme(
-          data: ThemeData.light().copyWith(
-            colorScheme: ColorScheme.light(primary: Color(0xFF4454C3)),
-            textButtonTheme: TextButtonThemeData(
-                style: TextButton.styleFrom(primary: Color(0xFF4454C3))),
-          ),
-          child: child!,
-        );
-      },
-    );
-
-    if (picked != null) {
-      final formattedDate = DateFormat.yMMMMd().format(picked);
-      birthdateController.text = formattedDate;
+      print('Error: $error');
+      // Return a default value or throw an exception if needed
+      return 0;
     }
   }
 
@@ -100,7 +87,7 @@ class _RegistrationState extends State<Registration> {
                       alignment: Alignment
                           .centerLeft, // Align the content to the left within the container
                       child: const Text(
-                        "Create your Account",
+                        "Login to your Account",
                         style: TextStyle(
                           color: Color(0xFF424E79),
                           fontSize: 17,
@@ -112,50 +99,57 @@ class _RegistrationState extends State<Registration> {
                       height: 15.0,
                     ),
                     CustomTextField(
-                      labelText: "Email Address",
-                      hintText: "Enter your Email Address",
-                      controller: emailController,
-                      textInputType: TextInputType.emailAddress,
-                    ),
-                    const SizedBox(
-                      height: 20.0,
-                    ),
-                    CustomTextField(
-                      labelText: "Phone Number",
-                      hintText: "0987 654 321",
-                      controller: phoneController,
-                      textInputType: TextInputType.phone,
+                      labelText: "Email or Mobile Number",
+                      hintText: "Enter your Email or Mobile Number",
+                      controller: inputController,
+                      textInputType: TextInputType.text,
                     ),
                     const SizedBox(
                       height: 20.0,
                     ),
                     PasswordField(
-                      obscureText: obscurePassword,
-                      onTap: handleObscurePassword,
-                      labelText: "Password",
-                      hintText: "Enter your Password",
-                      controller: passwordController,
+                        obscureText: obscurePassword,
+                        onTap: handleObscurePassword,
+                        labelText: "Password",
+                        hintText: "Enter your Password",
+                        controller: passwordController),
+                    const SizedBox(
+                      height: 5.0,
+                    ),
+                    Container(
+                      child: Row(
+                        children: [
+                          GestureDetector(
+                            onTap: () {},
+                            child: const Text(
+                              "Forgot your password?",
+                              style: TextStyle(
+                                color: Color(0xFF4454C3),
+                                fontSize: 15,
+                                fontWeight: FontWeight.normal,
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
                     ),
                     const SizedBox(
-                      height: 20.0,
-                    ),
-                    DatePickerField(
-                      labelText: "Birth Date",
-                      hintText: "Select your Birthday",
-                      inputWidth: width,
-                      inputHeight: 60,
-                      controller: birthdateController,
-                      function: () => _handleDateSelection(context),
-                    ),
-                    const SizedBox(
-                      height: 20.0,
+                      height: 25.0,
                     ),
                     PrimaryButton(
-                      text: "Sign up",
+                      text: "Login",
                       buttonColor: Color(0xFF4454C3),
-                      onPress: () {
-                        _RegisterPatient();
-                        Navigator.pushNamed(context, Login.routeName);
+                      onPress: () async {
+                        int parsedData = await _LoginPatient();
+                        if (parsedData != -1) {
+                          Navigator.pushNamed(context, Dashboard.routeName);
+                          globalId = parsedData;
+                          print(globalId);
+                        } else {
+                          // Redirect to login page or handle the situation accordingly
+                          Navigator.of(context).pushNamedAndRemoveUntil(
+                              Login.routeName, (Route<dynamic> route) => false);
+                        }
                       },
                     ),
                     const SizedBox(
@@ -205,7 +199,7 @@ class _RegistrationState extends State<Registration> {
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: [
                             const Text(
-                              "Already have an account?",
+                              "Don't have an account?",
                               style: TextStyle(
                                 color: Color(0xFF424E79),
                                 fontSize: 15,
@@ -213,10 +207,11 @@ class _RegistrationState extends State<Registration> {
                             ),
                             GestureDetector(
                               onTap: () {
-                                Navigator.pushNamed(context, Login.routeName);
+                                Navigator.pushNamed(
+                                    context, Registration.routeName);
                               },
                               child: const Text(
-                                " Login",
+                                " Sign up",
                                 style: TextStyle(
                                   color: Color(0xFF4454C3),
                                   fontSize: 15,
@@ -242,9 +237,5 @@ class _RegistrationState extends State<Registration> {
     setState(() {
       obscurePassword = !obscurePassword;
     });
-  }
-
-  Future<void> _handleDateSelection(BuildContext context) async {
-    await _selectDate(context);
   }
 }
